pyproject.toml
[project]
name = "copilot-directline-oauth"
version = "0.1.0"
description = "Secure Python solution for accessing Copilot Studio agents via Direct Line API with Entra ID OAuth authentication"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "msal>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
]

[project.scripts]
copilot-chat = "src.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build]
packages = ["src/copilot_directline", "src/cli"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py39"]
